Rename citizenC100CaseUtils.ts to citizenC100Case.utils.ts
