+++ e2e/journeys/citizen/caseView/uploadDocuments/witnessStatementsAndEvidence/applicant/uploadDocumentsPositionStatement.ts
@@ -12,7 +12,7 @@ import { DocumentSubmittedPage } from "../../../../../../pages/citizen/caseView/
 import {
   CitizenC100CaseUtils,
   CitizenUploadedDocument,
-} from "../../../../../../utils/citizenC100CaseUtils.ts";
+} from "../../../../../../utils/citizenC100Case.utils.ts";
 
 interface uploadDocumentsPositionStatementParams {
   page: Page;
